rename logic dir
