+++ apps/ledger-live-desktop/src/renderer/components/PostOnboardingHub/logic/useNavigateToPostOnboardingHubCallback.ts
